% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials in each experiment}

\item{p}{Probability of success on each trial}
}
\value{
A vector of simulated proportions for 0 through n successes
}
\description{
Binomial simulation
}
\examples{
\dontrun{mybin(iter = 1000,n = 10, p = 0.7)}
}
