% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal curve shading and probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the distribution}

\item{sigma}{Standard  deviation of the distribution}

\item{a}{The x-value to calculate P(X <= a)}
}
\value{
A list containing mu, sigma, and the calculated probability
}
\description{
Normal curve shading and probability
}
\examples{
\dontrun{myncurve(mu = 10, sigma = 5, a = 6)}
}
