% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Optimal number of airline tickets to sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the aircraft}

\item{gamma}{Maximum acceptable probability that the flight is overbooked}

\item{p}{Probability that a ticketed passenger shows up}
}
\value{
A named list containing nd (optimal tickets through discrete binomial), nc (optimal tickets through normal approximation), N, p, and gamma
}
\description{
Optimal number of airline tickets to sell
}
\examples{
\dontrun{ntickets(N = 400, gamma = 0.02, p = 0.95)}
}
